refactor(agent-chat): split tool presenters into one component per file
Move ToolMessageTiming, ToolInputDetails, ToolJsonDetails,
WorkflowToolMessage, RegularToolSummary, and RegularToolMessage into
their own files. Keep the shared tool icon in agent-chat-tool-icon.tsx.

Update the message body, the Computer Use card, and the tool input
details test.

This clears the react-doctor no-multi-component-file warnings that
block the CI check.
